+++ tests/test-scroll-grid.c
@@ -9,7 +9,7 @@ stage_size_notify_cb (ClutterActor *stage,
                       GParamSpec *pspec,
                       ClutterActor *table)
 {
-  guint width, height;
+  gfloat width, height;
 
   clutter_actor_get_size (stage, &width, &height);
   clutter_actor_set_size (table, width, height);
